requirements.txt
Flask
gunicorn
Flask-Dance[sqla]
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
psycopg2-binary
python-dotenv # Useful for managing environment variables locally
